refactor: Change util package name to constant
