+++ src/main/java/codesquad/constant/Message.java
@@ -1,4 +1,4 @@
-package codesquad.util;
+package codesquad.constant;
 
 public class Message {
     public static final String UNAUTHENTICATED = "UnAuthentication";
