+++ src/main/java/codesquad/user/dto/UserSignupDto.java
@@ -3,7 +3,7 @@ package codesquad.user.dto;
 
 import codesquad.user.domain.Role;
 import codesquad.user.domain.User;
-import codesquad.util.Regex;
+import codesquad.constant.Regex;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
